Add project-factory preconditions checker script (#27)
The Project Factory requires a number of enabled APIs and granted
permissions on the service account. The absence of these requirements
can cause the Project Factory to fail to run, generate invalid projects,
and fail in difficult to debug ways.

This commit adds a python script to validate that the Project Factory
preconditions are met, based on the configuration given to the script.