Make `credentials_path` optional, add auth probing
When Terraform is run on GCE instances, Terraform can use service
account credentials associated with the instance. However the
project-factory module required that a value for `credentials_path` in
order to delete the default GCE service account. This increases the
barrier to entry for the project-factory.

This commit makes `credentials_path` optional by falling back to the
active gcloud credentials when service account credentials aren't
provided. When Terraform runs, it will validate that there are
credentials for deleting the GCE service account - preferring
`credentials_path` and then falling back to the active gcloud
credentials.