Create uniquely named shared vpc in full test suite
Tests fail if this network already exists in the targeted project.

Using unique network names allows the creation of test networks in the
same project for all running tests.

This protects against race conditions should many instances of the suite
be running in parallel. As well if a test suite was not fully cleaned up
(destroyed).